pyproject.toml
[project]
name = "google-workspace-orchestrator"
version = "0.1.0"
description = "Intelligent orchestrator for Gmail, Google Calendar, and Google Drive using natural language"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "authlib>=1.6.8",
    "celery>=5.6.2",
    "cryptography>=46.0.5",
    "fastapi>=0.131.0",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
